refactor(views): flatten project dir structure
