+++ src/views/project/components/EDAProject.vue
