+++ src/views/project/components/EDATargetNextpnr.vue
