+++ src/views/project/components/EDATargetSynthesis.vue
