+++ src/views/project/components/EDATargetYosys.vue
