--- src/views/project/index.html
@@ -1,1 +1,0 @@
-../_common/index.html
